+++ projects/ids-enterprise-ng/src/lib/lookup/soho-lookup.component.ts
@@ -16,7 +16,7 @@ import {
 import {
   BaseControlValueAccessor,
   provideControlValueAccessor
-} from '../utils/base-control-value-accessor';
+} from '../utils';
 
 @Component({
   selector: 'input[soho-lookup]', // eslint-disable-line
@@ -53,8 +53,8 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
     }
 
     this._dataset = data;
-    if (data && this.jQueryElement && this.lookup?.settings) {
-      (this.lookup as any).settings.options.dataset = data;
+    if (data && this.jQueryElement && this.lookup?.settings?.options) {
+      this.lookup.settings.options.dataset = data;
       this.markForUpdate();
     }
   }
@@ -361,18 +361,22 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
     return true;
   }
 
-  @HostBinding('attr.disabled') get isDisabledAttr() {
-    return this.isDisabled || undefined;
-  }
-
+  /**
+   * Disables ths lookup control.
+   */
+  @HostBinding('disabled')
   @Input() isDisabled: boolean | undefined = undefined;
-
 
   /**
    * Local variables
    */
   private jQueryElement?: JQuery;
 
+  /**
+   * Soho Lookup widget.
+   *
+   * @private
+   */
   private lookup?: SohoLookupStatic | null;
 
   private settings: SohoLookupOptions = {};
@@ -380,6 +384,13 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
   /** Initial dataset */
   private _dataset?: Object[];
 
+
+  /**
+   * Constructor.
+   *
+   * @param element
+   * @param ngZone
+   */
   constructor(private element: ElementRef, private ngZone: NgZone) {
     super();
   }
@@ -457,16 +468,14 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
   /** Enable the input. **/
   public enable(): void {
-    if (this.lookup) {
-      this.ngZone.runOutsideAngular(() => this.lookup?.enable());
-    }
+    this.isDisabled = false;
+    this.markForUpdate();
   }
 
   /** Disable the input. **/
   public disable(): void {
-    if (this.lookup) {
-      this.ngZone.runOutsideAngular(() => this.lookup?.disable());
-    }
+    this.isDisabled = true;
+    this.markForUpdate();
   }
 
   /** Make the input readonly. **/
@@ -479,7 +488,7 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
   /** Input is disabled or not **/
   public isReadonly(): boolean | undefined {
     if (this.lookup) {
-      return this.ngZone.runOutsideAngular(() => (this.lookup as any).isReadonly());
+      return this.ngZone.runOutsideAngular(() => this.lookup?.isReadonly());
     }
   }
 
@@ -606,7 +615,7 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
       if (typeof toProcess[i] === 'object') {
         if (typeof this.settings.field === 'function') {
-          current = (this.settings.field as SohoLookupFieldFunction)(toProcess[i], (this.lookup as any).element, (this.lookup as any).grid);
+          current = (this.settings.field as SohoLookupFieldFunction)(toProcess[i], this.lookup!.element, this.lookup!.grid);
         } else {
           current = (toProcess[i] as any)[this.settings.field as string];
         }
@@ -667,9 +676,9 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
     }
 
     if (event.length && event.length === 1 && !this.isMultiselect()) {
-      this.internalValue = this.asobject !== false ? event[0].data : this.processValue(event[0].data);
+      this.internalValue = this.asobject ? event[0].data : this.processValue(event[0].data);
     } else {
-      this.internalValue = event.map(val => this.asobject !== false ? val.data : this.processValue(val.data));
+      this.internalValue = event.map(val => this.asobject ? val.data : this.processValue(val.data));
     }
   }