+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 10.2.0 Fixes
 
+- `[Lookup]` Fixed a bug where the enabling the lookup doesn't work when the component is initially disabled. ([#1069](https://github.com/infor-design/enterprise-ng/issues/1069))
 - `[Multiselect]` Added an example showing automation id attributes on options in multiselect. ([#1005](https://github.com/infor-design/enterprise-ng/issues/1005))
 - `[SohoMessageService]` Removed `undefined` return type from messages method. ([#1061](https://github.com/infor-design/enterprise-ng/issues/1061))
 - `[SohoDatePicker]` Fixed datepicker format when assigned via angular form control. ([#1072](https://github.com/infor-design/enterprise-ng/issues/1072))